+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/tests/latest/test_acr_commands_mock.py
@@ -53,14 +53,12 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.repository.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_repository_list(self, mock_requests_get, mock_get_access_credentials, mock_get_registry_by_name):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
-        encoded_repositories = json.dumps({'repositories': ['testrepo1', 'testrepo2']}).encode()
+        cmd = self._setup_cmd()
 
         response = mock.MagicMock()
         response.headers = {}
         response.status_code = 200
-        response.content = encoded_repositories
+        response.content = json.dumps({'repositories': ['testrepo1', 'testrepo2']}).encode()
         mock_requests_get.return_value = response
 
         # List repositories using Basic auth on a classic registry
@@ -97,8 +95,8 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.repository.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_repository_show_tags(self, mock_requests_get, mock_get_access_credentials, mock_get_registry_by_name):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
+        cmd = self._setup_cmd()
+
         encoded_tags = json.dumps({'tags': ['testtag1', 'testtag2']}).encode()
         encoded_tags_detail = json.dumps({'tags': [
             {
@@ -154,9 +152,12 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.repository.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_repository_show_manifests(self, mock_requests_get, mock_get_access_credentials, mock_get_registry_by_name):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
-        encoded_manifests = json.dumps({'manifests': [
+        cmd = self._setup_cmd()
+
+        response = mock.MagicMock()
+        response.headers = {}
+        response.status_code = 200
+        response.content = json.dumps({'manifests': [
             {
                 'digest': 'sha256:b972dda797ef258a7ea5738eb2109778c2bac6a99d1033e6c9f9bdb4fbd196e7',
                 'tags': ['testtag1', 'testtag2']
@@ -165,11 +166,6 @@ class AcrMockCommandsTests(unittest.TestCase):
                 'digest': 'sha256:c5515758d4c5e1e838e9cd307f6c6a0d620b5e07e6f927b07d05f6d12a1ac8d7',
                 'tags': ['testtag3']
             }]}).encode()
-
-        response = mock.MagicMock()
-        response.headers = {}
-        response.status_code = 200
-        response.content = encoded_manifests
         mock_requests_get.return_value = response
 
         # Show manifests using Basic auth without detail
@@ -208,17 +204,15 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.repository.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_repository_show(self, mock_requests_get, mock_get_access_credentials, mock_get_registry_by_name):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
-        encoded_manifests = json.dumps({
-            'registry': 'testregistry.azurecr.io',
-            'imageName': 'testrepository'
-        }).encode()
+        cmd = self._setup_cmd()
 
         response = mock.MagicMock()
         response.headers = {}
         response.status_code = 200
-        response.content = encoded_manifests
+        response.content = json.dumps({
+            'registry': 'testregistry.azurecr.io',
+            'imageName': 'testrepository'
+        }).encode()
         mock_requests_get.return_value = response
 
         mock_get_registry_by_name.return_value = Registry(location='westus', sku=Sku(name='Standard')), 'testrg'
@@ -265,8 +259,7 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('requests.request', autospec=True)
     @mock.patch('requests.get', autospec=True)
     def test_repository_delete(self, mock_requests_get, mock_requests_delete, mock_get_access_credentials, mock_get_registry_by_name):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
+        cmd = self._setup_cmd()
 
         get_response = mock.MagicMock()
         get_response.headers = {
@@ -340,42 +333,6 @@ class AcrMockCommandsTests(unittest.TestCase):
             json=None,
             verify=mock.ANY)
 
-        # Delete tag (deprecating)
-        acr_repository_delete(cmd, 'testregistry', 'testrepository', tag='testtag', yes=True)
-        mock_requests_delete.assert_called_with(
-            method='delete',
-            url='https://testregistry.azurecr.io/v2/_acr/testrepository/tags/testtag',
-            headers=get_authorization_header('username', 'password'),
-            params=None,
-            json=None,
-            verify=mock.ANY)
-
-        # Delete manifest with tag (deprecating)
-        acr_repository_delete(cmd, 'testregistry', 'testrepository', tag='testtag', manifest='', yes=True)
-        expected_get_headers = get_authorization_header('username', 'password')
-        expected_get_headers.update(MANIFEST_V2_HEADER)
-        mock_requests_get.assert_called_with(
-            url='https://testregistry.azurecr.io/v2/testrepository/manifests/testtag',
-            headers=expected_get_headers,
-            verify=mock.ANY)
-        mock_requests_delete.assert_called_with(
-            method='delete',
-            url='https://testregistry.azurecr.io/v2/testrepository/manifests/sha256:c5515758d4c5e1e838e9cd307f6c6a0d620b5e07e6f927b07d05f6d12a1ac8d7',
-            headers=get_authorization_header('username', 'password'),
-            params=None,
-            json=None,
-            verify=mock.ANY)
-
-        # Delete manifest with digest (deprecating)
-        acr_repository_delete(cmd, 'testregistry', 'testrepository', manifest='sha256:c5515758d4c5e1e838e9cd307f6c6a0d620b5e07e6f927b07d05f6d12a1ac8d7', yes=True)
-        mock_requests_delete.assert_called_with(
-            method='delete',
-            url='https://testregistry.azurecr.io/v2/testrepository/manifests/sha256:c5515758d4c5e1e838e9cd307f6c6a0d620b5e07e6f927b07d05f6d12a1ac8d7',
-            headers=get_authorization_header('username', 'password'),
-            params=None,
-            json=None,
-            verify=mock.ANY)
-
     @mock.patch('azure.cli.command_modules.acr._docker_utils.get_registry_by_name', autospec=True)
     @mock.patch('requests.post', autospec=True)
     @mock.patch('requests.get', autospec=True)
@@ -427,8 +384,7 @@ class AcrMockCommandsTests(unittest.TestCase):
                                    tenant_suffix=test_tenant_suffix)
 
     def _core_token_scenarios(self, mock_get_raw_token, mock_requests_get, mock_requests_post, mock_get_registry_by_name, registry_exists, registry_name, login_server, tenant_suffix):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
+        cmd = self._setup_cmd()
 
         if registry_exists:
             registry = Registry(location='westus', sku=Sku(name='Standard'))
@@ -441,15 +397,15 @@ class AcrMockCommandsTests(unittest.TestCase):
         self._setup_mock_token_requests(mock_get_raw_token, mock_requests_get, mock_requests_post, login_server)
 
         # Test get refresh token
-        get_login_credentials(cmd.cli_ctx, registry_name, tenant_suffix=tenant_suffix)
+        get_login_credentials(cmd, registry_name, tenant_suffix=tenant_suffix)
         self._validate_refresh_token_request(mock_requests_get, mock_requests_post, login_server)
 
         # Test get access token for container image repository
-        get_access_credentials(cmd.cli_ctx, registry_name, tenant_suffix=tenant_suffix, repository=TEST_REPOSITORY, permission='*')
+        get_access_credentials(cmd, registry_name, tenant_suffix=tenant_suffix, repository=TEST_REPOSITORY, permission='*')
         self._validate_access_token_request(mock_requests_get, mock_requests_post, login_server, 'repository:{}:*'.format(TEST_REPOSITORY))
 
         # Test get access token for artifact image repository
-        get_access_credentials(cmd.cli_ctx, registry_name, tenant_suffix=tenant_suffix, artifact_repository=TEST_REPOSITORY, permission='*')
+        get_access_credentials(cmd, registry_name, tenant_suffix=tenant_suffix, artifact_repository=TEST_REPOSITORY, permission='*')
         self._validate_access_token_request(mock_requests_get, mock_requests_post, login_server, 'artifact-repository:{}:*'.format(TEST_REPOSITORY))
 
     def _setup_mock_token_requests(self, mock_get_aad_token, mock_requests_get, mock_requests_post, login_server):
@@ -501,9 +457,12 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.helm.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_helm_list(self, mock_requests_get, mock_get_access_credentials):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
-        encoded_charts = json.dumps({
+        cmd = self._setup_cmd()
+
+        response = mock.MagicMock()
+        response.headers = {}
+        response.status_code = 200
+        response.content = json.dumps({
             'mychart1': [
                 {
                     'name': 'mychart1',
@@ -520,11 +479,6 @@ class AcrMockCommandsTests(unittest.TestCase):
                     'version': '2.1.0'
                 }
             ]}).encode()
-
-        response = mock.MagicMock()
-        response.headers = {}
-        response.status_code = 200
-        response.content = encoded_charts
         mock_requests_get.return_value = response
 
         mock_get_access_credentials.return_value = 'testregistry.azurecr.io', EMPTY_GUID, 'password'
@@ -540,9 +494,12 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.helm.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_helm_show(self, mock_requests_get, mock_get_access_credentials):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
-        encoded_charts = json.dumps({
+        cmd = self._setup_cmd()
+
+        response = mock.MagicMock()
+        response.headers = {}
+        response.status_code = 200
+        response.content = json.dumps({
             'mychart1': [
                 {
                     'name': 'mychart1',
@@ -553,11 +510,6 @@ class AcrMockCommandsTests(unittest.TestCase):
                     'version': '0.1.2'
                 }
             ]}).encode()
-
-        response = mock.MagicMock()
-        response.headers = {}
-        response.status_code = 200
-        response.content = encoded_charts
         mock_requests_get.return_value = response
 
         mock_get_access_credentials.return_value = 'testregistry.azurecr.io', EMPTY_GUID, 'password'
@@ -585,8 +537,7 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.helm.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_helm_delete(self, mock_requests_get, mock_get_access_credentials):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
+        cmd = self._setup_cmd()
 
         response = mock.MagicMock()
         response.headers = {}
@@ -618,8 +569,7 @@ class AcrMockCommandsTests(unittest.TestCase):
     @mock.patch('azure.cli.command_modules.acr.helm.get_access_credentials', autospec=True)
     @mock.patch('requests.request', autospec=True)
     def test_helm_push(self, mock_requests_get, mock_get_access_credentials):
-        cmd = mock.MagicMock()
-        cmd.cli_ctx = DummyCli()
+        cmd = self._setup_cmd()
 
         response = mock.MagicMock()
         response.headers = {}
@@ -665,3 +615,14 @@ class AcrMockCommandsTests(unittest.TestCase):
                 params=None,
                 data=mock_open.return_value.__enter__.return_value,
                 verify=mock.ANY)
+
+    def _setup_cmd(self):
+        cmd = mock.MagicMock()
+        cmd.cli_ctx = DummyCli()
+        mock_sku = mock.MagicMock()
+        mock_sku.classic.value = 'Classic'
+        mock_sku.basic.value = 'Basic'
+        mock_sku.standard.value = 'Standard'
+        mock_sku.premium.value = 'Premium'
+        cmd.get_models.return_value = mock_sku
+        return cmd