+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -7,6 +7,8 @@ Release History
 ++++++
 * Add support for VNet/IP rules.
 * Add --platform parameter for `acr build`, `acr task create` and `acr task update` commands
+* BREAKING CHANGE: Remove 'acr build-task' command group.
+* BREAKING CHANGE: Remove --tag/--manifest from 'acr repository delete' command.
 
 2.1.12
 ++++++